requirements.txt
fastapi[gunicorn]

pydantic
pydantic-settings

pyjwt